+++ mycli/main.py
@@ -509,8 +509,7 @@ class MyCli(object):
                 start = time()
                 res = sqlexecute.run(document.text)
                 successful = True
-                output = []
-                total = 0
+                result_count = 0
                 for title, cur, headers, status in res:
                     logger.debug("headers: %r", headers)
                     logger.debug("rows: %r", cur)
@@ -533,9 +532,20 @@ class MyCli(object):
                                                    special.is_expanded_output(),
                                                    max_width)
 
-                    output.extend(formatted)
-                    total = time() - start
+                    t = time() - start
+                    try:
+                        if result_count > 0:
+                            self.echo('')
+                        self.output('\n'.join(formatted))
+                        if special.is_timing_enabled():
+                            self.echo('Time: %0.03fs' % t)
+                    except KeyboardInterrupt:
+                        pass
+
+                    start = time()
+                    result_count += 1
                     mutating = mutating or is_mutating(status)
+                special.unset_once_if_written()
             except EOFError as e:
                 raise e
             except KeyboardInterrupt:
@@ -580,13 +590,6 @@ class MyCli(object):
                 logger.error("traceback: %r", traceback.format_exc())
                 self.echo(str(e), err=True, fg='red')
             else:
-                try:
-                    self.output('\n'.join(output))
-                except KeyboardInterrupt:
-                    pass
-                if special.is_timing_enabled():
-                    self.echo('Time: %0.03fs' % total)
-
                 # Refresh the table names and column names if necessary.
                 if need_completion_refresh(document.text):
                     self.refresh_completions(