+++ mycli/sqlexecute.py
@@ -123,31 +123,40 @@ class SQLExecute(object):
             if sql.endswith('\\G'):
                 special.set_expanded_output(True)
                 sql = sql[:-2].strip()
+
+            cur = self.conn.cursor()
             try:   # Special command
                 _logger.debug('Trying a dbspecial command. sql: %r', sql)
-                cur = self.conn.cursor()
                 for result in special.execute(cur, sql):
                     yield result
             except special.CommandNotFound:  # Regular SQL
-                yield self.execute_normal_sql(sql)
+                _logger.debug('Regular sql statement. sql: %r', sql)
+                cur.execute(sql)
+                while True:
+                    yield self.get_result(cur)
 
-    def execute_normal_sql(self, split_sql):
-        _logger.debug('Regular sql statement. sql: %r', split_sql)
-        cur = self.conn.cursor()
-        num_rows = cur.execute(split_sql)
+                    # PyMySQL returns an extra, empty result set with stored
+                    # procedures. We skip it (rowcount is zero and no
+                    # description).
+                    if not cur.nextset() or (not cur.rowcount and cur.description is None):
+                        break
+
+    def get_result(self, cursor):
+        """Get the current result's data from the cursor."""
         title = headers = None
 
-        # cur.description is not None for queries that return result sets, e.g.
-        # SELECT or SHOW.
-        if cur.description is not None:
-            headers = [x[0] for x in cur.description]
+        # cursor.description is not None for queries that return result sets,
+        # e.g. SELECT or SHOW.
+        if cursor.description is not None:
+            headers = [x[0] for x in cursor.description]
             status = '{0} row{1} in set'
         else:
             _logger.debug('No rows in result.')
             status = 'Query OK, {0} row{1} affected'
-        status = status.format(num_rows, '' if num_rows == 1 else 's')
+        status = status.format(cursor.rowcount,
+                               '' if cursor.rowcount == 1 else 's')
 
-        return (title, cur if cur.description else None, headers, status)
+        return (title, cursor if cursor.description else None, headers, status)
 
     def tables(self):
         """Yields table names"""